Rename bench-gif-dot-rs to bench-rust-gif-dot-rs
There may be future bench-go-xxx programs, or of other languages.

On my machine, the throughput number that the program output also
changes from 24.7 (as reported in a prior commit message) to 28.0
megapixels per second. This is not due to any code changes, but because
in the meantime, I have also used the "cpupower" command line tool to
change policy from "powersave" to "performance".