Rework how we avoid clang's -Wunused-function
The previous technique to avoid this was to have explicit

(void)(wuffs_foo__bar_baz);

no-op lines of code. However, on MSVC, this triggers warning C4551
"function call missing argument list".

https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-1-c4551
says that "A function call must include the open and close parentheses".
But this is not a function call, it is casting the function pointer.
This false positive is a bug in MSVC, but we work around it for Wuffs.