Replace RAC's Brotli Short Codec value by LZ4
RAC still supports Brotli (as a Long Codec, not a Short Codec). It's
just no longer given a dedicated Short Codec value (and therefore less
likely to be supported out-of-the-box by cmd/ractool). Brotli occupies a
similar position to Zstandard on the speed versus size frontier. LZ4
occupies a noticably different position, faster but larger.